converted lil matrix to float

+++ stellarscope/stellarscope_assign.py
@@ -50,7 +50,7 @@ def fit_telescope_model(ts: Stellarscope, pooling_mode: str) -> TelescopeLikelih
     if pooling_mode == 'individual':
 
         ''' Initialise the z matrix for all reads '''
-        z = lil_matrix(ts.raw_scores)
+        z = lil_matrix(ts.raw_scores, dtype=np.float64)
         for barcode in ts.barcodes:
             if barcode in ts.barcode_read_indices:
                 _rows = ts.barcode_read_indices[barcode]
